backend/pyproject.toml
[project]
name = "hostelive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==4.2",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "google-generativeai>=0.8.5",
    "pillow>=11.2.1",
    "vadersentiment>=3.3.2",
]
